made some utils files, put part of the refresh data code in the utils file
